Move caps.Repository to a dedicated file
